// src/Frontend/ViteApp/src/api-generated/models/LayoutCreateRequestModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KISv4 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { LayoutItemCreateRequest } from './LayoutItemCreateRequest';
import {
    LayoutItemCreateRequestFromJSON,
    LayoutItemCreateRequestFromJSONTyped,
    LayoutItemCreateRequestToJSON,
    LayoutItemCreateRequestToJSONTyped,
} from './LayoutItemCreateRequest';

/**
 * 
 * @export
 * @interface LayoutCreateRequestModel
 */
export interface LayoutCreateRequestModel {
    /**
     * 
     * @type {string}
     * @memberof LayoutCreateRequestModel
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof LayoutCreateRequestModel
     */
    image?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof LayoutCreateRequestModel
     */
    topLevel?: boolean;
    /**
     * 
     * @type {Array<LayoutItemCreateRequest>}
     * @memberof LayoutCreateRequestModel
     */
    layoutItems: Array<LayoutItemCreateRequest>;
}

/**
 * Check if a given object implements the LayoutCreateRequestModel interface.
 */
export function instanceOfLayoutCreateRequestModel(value: object): value is LayoutCreateRequestModel {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('layoutItems' in value) || value['layoutItems'] === undefined) return false;
    return true;
}

export function LayoutCreateRequestModelFromJSON(json: any): LayoutCreateRequestModel {
    return LayoutCreateRequestModelFromJSONTyped(json, false);
}

export function LayoutCreateRequestModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): LayoutCreateRequestModel {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'image': json['image'] == null ? undefined : json['image'],
        'topLevel': json['topLevel'] == null ? undefined : json['topLevel'],
        'layoutItems': ((json['layoutItems'] as Array<any>).map(LayoutItemCreateRequestFromJSON)),
    };
}

export function LayoutCreateRequestModelToJSON(json: any): LayoutCreateRequestModel {
    return LayoutCreateRequestModelToJSONTyped(json, false);
}

export function LayoutCreateRequestModelToJSONTyped(value?: LayoutCreateRequestModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'image': value['image'],
        'topLevel': value['topLevel'],
        'layoutItems': ((value['layoutItems'] as Array<any>).map(LayoutItemCreateRequestToJSON)),
    };
}